Fix matching for incomplete CUSD answers
The Sony Ericsson modem in an IdeaPad S10 3G answers CUSD commands
in cleartext without an encoding (third argument of CUSD answer).
The regular expression will now match and the script will assume
a cleartext answer if no encoding is sent in the answer.

Thanks to Ulf (yeti@debianforum.de) for providing a debug output
of gsm-ussd to diagnose this problem!